Week2/Daily Project/Story_Generator/backend/requirements.txt
Flask
langchain
langchain-google-genai
google-generativeai
flask-cors
Werkzeug==2.3.8